GroceryAppDemo/backend/requirements.txt
Flask==2.3.3
elasticsearch==8.11.0
openai==1.3.0
anthropic==0.7.0
python-dotenv==1.0.0
